Refactor of Lesson 1 and Lesson2 variables//
