+++ src/main/java/teacher/students/student_Mihails_Cimotiss/Lesson_2/variablesPrimitive.java
@@ -1,0 +1,29 @@
+package main.java.teacher.students.student_Mihails_Cimotiss.Lesson_2;
+public class variablesPrimitive {
+    public static void main(String[] args) {
+        byte myBite;
+        myBite = 8;
+        System.out.println("myByte value is:" + myBite);
+        int myInt;
+        myInt = 32000;
+        System.out.println("myByte value is:"  +myInt);
+        short myShort = 100;
+        System.out.println("myByte value is:" + myShort);
+
+        long myLong = 1_000_000;
+        System.out.println("myByte value is:" + myLong);
+
+        float myFloat = 3.14f;
+        double myDouble = 3.54;
+        System.out.println( "myFloat value is:" +myFloat);
+
+        boolean approve = true;
+        boolean decline = false;
+        System.out.println("boolean:");
+
+        char myChar='B';
+        System.out.println ("char is not needed:");
+
+    }
+}
+
